+++ src/components/ConnectedTasksJobs/ConnectedTasksJobsPane.js
@@ -1,0 +1,12 @@
+import { ConnectedTasksJobsPlugin, connectedTasksJobsPropTypes } from './ConnectedTasksJobsPlugin';
+
+const ConnectedTasksJobsPane = props => (
+  <ConnectedTasksJobsPlugin
+    {...props}
+    componentType="ConnectedTasksJobsPane"
+  />
+);
+
+ConnectedTasksJobsPane.propTypes = connectedTasksJobsPropTypes;
+
+export default ConnectedTasksJobsPane;